requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
polars>=0.20.0
duckdb>=0.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
dtaidistance>=2.3.0
numpy>=1.26.0
